Restore Snap Support
Alter cmake file and snapcraft.yaml to build snap builds once more. Due to core22
not supporting qmake yet [1] and Qt not continuing to support qmake, this commit moves
Packet Sender snap builds to cmake. In order to not mess with Debian builds (Which use
the existing CMakeLists.txt), a variable SNAP has to be defined and set to a truthy value
in order to do snap-specific changes.

The alterations were generated by using qmake2cmake [3] on the .pro file for snap builds,
and then merged under if statements for the existing CMakeLists.txt. I did have to comment out
the git sha generation and add that back manually for the tool to work

[1] https://forum.snapcraft.io/t/core22-qmake-plugin-support/35026
[2] https://www.qt.io/blog/2019/08/07/technical-vision-qt-6
[3] https://code.qt.io/cgit/qt/qmake2cmake.git/about/